requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==5.0.7
django-crispy-forms==2.2
django-debug-toolbar==4.4.2
django-filter==24.2
djangorestframework==3.15.2
pillow==10.4.0
pytz==2023.3.post1
sqlparse==0.5.0
typing_extensions==4.9.0
tzdata==2023.4
